refactor: Move LoggingRetry to its own eth_defi.logging_retry submodule
Move the reusable LoggingRetry HTTP retry adapter out of the Velvet
submodule into a top-level eth_defi.logging_retry module. It is shared
across the Velvet, Derive, Core3, Hibachi, GRVT, Lighter, Hyperliquid
and token-analysis integrations, so nesting it under velvet was
misleading. Update all imports and relocate the Sphinx autosummary
entry from the Velvet API index to the Core API index.
