+++ docs/source/api/velvet/index.rst
@@ -13,4 +13,3 @@ Velvet Capital vaults integration.
    eth_defi.velvet.redeem
    eth_defi.velvet.enso
    eth_defi.velvet.analysis
-   eth_defi.velvet.logging_retry
