+++ eth_defi/velvet/enso.py
@@ -13,7 +13,7 @@ from requests.exceptions import RetryError
 from requests.sessions import HTTPAdapter
 
 from eth_defi.velvet.config import VELVET_DEFAULT_API_URL, VELVET_GAS_EXTRA_SAFETY_MARGIN
-from eth_defi.velvet.logging_retry import LoggingRetry
+from eth_defi.logging_retry import LoggingRetry
 
 logger = logging.getLogger(__name__)
